pyproject.toml
[tool.poetry]
name = "repl_python3_WordPacks"
version = "0.1.0"
description = ""
authors = ["diogeek <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.3"
python-dotenv = "^0.20.0"
discord = "^1.7.3"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
